Breaks utils module up into seperate submodules
