+++ sndata/utils/__init__.py
@@ -1,0 +1,3 @@
+"""The ``utils`` module provides an assorted collection of general utilities
+used when building data access classes for a given survey / data release.
+"""
